+++ Something/WebApplication/wwwroot/js/apps/authentication/Authentication.scss
@@ -7,11 +7,4 @@
         height: 100%;
         padding: 10% 10px 10px 0px;
     }
-    .login, .register {
-        width: 400px;
-        margin: 0 auto;
-        h2 {
-            font-size: 32px;
-        }
-    }
 }
