+++ Something/WebApplication/wwwroot/js/apps/authentication/Authentication.tsx
@@ -1,11 +1,14 @@
-import * as React from 'react';
+import * as React from "react";
 
-import { LoginForm } from './components/LoginForm';
-import { Splash } from './components/Splash'
+import { LoginForm } from "./components/LoginForm/LoginForm";
+import { Splash } from "./components/Splash/Splash";
+
+import "./../Global.scss";
+import "./Authentication.scss";
 
 export default class Login extends React.Component<{}, {}> {
-    render() {
-        return <div className='authentication ant-row'>
+    render (): JSX.Element {
+        return <div className="authentication ant-row">
             <LoginForm />
             <Splash />
         </div>;
