+++ Something/WebApplication/wwwroot/js/apps/authentication/components/LoginForm/LoginForm.scss
@@ -1,0 +1,7 @@
+.LoginForm{
+    width: 400px;
+    margin: 0 auto;
+    h2 {
+        font-size: 32px;
+    }
+}
