--- Something/WebApplication/wwwroot/js/apps/authentication/components/Splash.tsx
@@ -1,14 +1,0 @@
-import * as React from 'react';
-
-export class Splash extends React.Component {
-    render() {
-        return (
-            <div className="pane splash ant-col-md-12">
-                <div className="splash--inner">
-                    <div className="splash--header">Explore Something!</div>
-                    <img src="https://cdn.dribbble.com/users/250027/screenshots/4065967/gifmaker.org_gll2rd.gif" />
-                </div>
-            </div>
-        );
-    }
-}
